pyproject.toml
[project]
name = "RR-GCL"
version = "0.0.1"
description = "Mutation-Guided Graph Contrastive Learning for Histone-related Residue Representation"
requires-python = ">=3.11"

authors = [
    {name = "Sunghwan Moon"},
    {name = "Yiru Sheng"},
    {name = "Anna Panchenko"}
]

dependencies = [
    "torch==2.7.1",
    "torch-geometric==2.7.0",
    "torch-sparse==0.6.18+torch27.computecanada",
    "torch-scatter==2.1.2+torch27.computecanada",
    "networkx==3.5",
    "pandas",
    "numpy",
    "scipy",
    "pyyaml",
    "wandb==0.24.0",
    "scikit-learn",
    "matplotlib",
    "statsmodels",
    "bioservices",
    "nx-parallel==0.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*", "models*", "data*"]